Working through sample expressions ... testing that they function as expected or revise what is expected by how they behave?
I've just got the github:// repositories working the way that I expect them to work, but I have to revisit the API and determine if things make sense. For example:

  Metacello new
    project: 'Sample';
    repository: 'github://dalehenrich/MetacelloRepository:master/repository';
    get.

loads the configuration from the master branch of the github project. But I am not sure what I want the following expression to mean:

  Metacello new
    project: 'Sample';
    version: #stable;
    repository: 'github://dalehenrich/MetacelloRepository:master/repository';
    load.

Or ... I think I know what I want it to mean, but I think that the two expressions are too close together to be useful.

Probably need to rethink the scripting API some more...